Refactor favourite functionality in details screen
This commit refactors the add/remove favourite functionality in the `DetailsScreen`.

- Renames `addPhotoToFavourite` and `remotePhotoFromFavourite` in `DetailsViewModel` to a single `togglePhotoFavourite` method.
- This new method checks the current favourite status and performs the appropriate add or remove action.
- Updates `DetailsScreen` to call `togglePhotoFavourite` when the favourite button is clicked.
- Changes the favourite button text to "Remove favourites" when an item is already a favourite, and "Add to favourites" otherwise.

+++ feature/details/src/main/java/com/segunfrancis/details/ui/DetailsViewModel.kt
@@ -73,15 +73,13 @@ class DetailsViewModel(
         }
     }
 
-    fun addPhotoToFavourite() {
+    fun togglePhotoFavourite() {
         viewModelScope.launch(exceptionHandler) {
-            photosResponseItem?.let { repository.addPhotoToFavourite(it) }
-        }
-    }
-
-    fun remotePhotoFromFavourite() {
-        viewModelScope.launch(exceptionHandler) {
-            photosResponseItem?.let { repository.removePhotoFromFavourite(it.id) }
+            if (uiState.value.isFavourite) {
+                photosResponseItem?.let { repository.removePhotoFromFavourite(it.id) }
+            } else {
+                photosResponseItem?.let { repository.addPhotoToFavourite(it) }
+            }
         }
     }
